LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/container-queries/anchored-fallback-position-area-expected.txt


FAIL @container anchored(fallback) with <position-area> fallback applied assert_equals: expected "yes" but got ""
